requirements.txt
# Telegram
python-telegram-bot>=20.0

# GitHub
PyGithub>=2.1.1

# Utils
loguru>=0.7.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0